requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
starlette==0.38.2
anyio==4.4.0
watchfiles==0.22.0
colorama==0.4.6
openai>=1.0.0
requests>=2.25.0
python-dotenv>=1.0.0
pylint>=3.0.0
flake8>=6.0.0
bandit>=1.7.0
mypy>=1.0.0
